Fixed Hard fault
I moved the adc value hash array to global variables. I suspect it's
16kB size was overflowing our stack.
I also initialized our time and adc variables manually, as the array we
were initializing from originally was not actually initialized at the
time of calling.
Not sure which bug fixed it, but the thing works now and thats all that
matters.